feat:custom role based permission check middleware change rebasE

+++ server/server.js
@@ -1029,7 +1029,10 @@ app.post("/api/TRByDateRange/", loginCheck, permissionCheck(Role("manager"),Role
 });
 
 
-app.post("/api/TR", loginCheck, permissionCheck(Role("manager"),Role("admin")), async function (req, res) {
+app.post("/api/TR", loginCheck, async function (req, res) {
+  if (req["user"]["ID"] === "guest") {
+    return res.send("게스트 계정은 저장, 수정, 삭제가 불가능합니다.");
+  }
   const newTR = req.body;
   // db.collection("TR").findOne({ ID: newTR.ID, 날짜: newTR.날짜 }, function (err, result) {
   //   if (err) {
@@ -1410,7 +1413,10 @@ app.get("/api/Closemeeting/:date", loginCheck, permissionCheck(Role("manager"),R
   });
 });
 
-app.put("/api/Closemeeting/:date", loginCheck, permissionCheck(Role("manager"),Role("admin")), function (req, res) {
+app.put("/api/Closemeeting/:date", loginCheck, function (req, res) {
+  if (req["user"]["ID"] === "guest") {
+    return res.send("게스트 계정은 저장, 수정, 삭제가 불가능합니다.");
+  }
   const paramDate = decodeURIComponent(req.params.date);
   const newClosemeeting = req.body;
   let findID;
@@ -1502,7 +1508,10 @@ app.get("/api/Middlemeeting/:date", loginCheck, permissionCheck(Role("manager"),
   });
 });
 
-app.put("/api/Middlemeeting/:date", loginCheck, permissionCheck(Role("manager"),Role("admin")), function (req, res) {
+app.put("/api/Middlemeeting/:date", loginCheck, function (req, res) {
+  if (req["user"]["ID"] === "guest") {
+    return res.send("게스트 계정은 저장, 수정, 삭제가 불가능합니다.");
+  }
   const paramDate = decodeURIComponent(req.params.date);
   const newMiddlemeeting = req.body;
   let findID;
@@ -1679,7 +1688,10 @@ app.post("/api/Textbook", loginCheck, permissionCheck(Role("manager"),Role("admi
   });
 });
 
-app.delete("/api/Textbook/:_id", loginCheck, permissionCheck(Role("manager"),Role("admin")), async (req, res) => {
+app.delete("/api/Textbook/:_id", loginCheck, async (req, res) => {
+  if (req["user"]["ID"] === "guest") {
+    return res.send("게스트 계정은 저장, 수정, 삭제가 불가능합니다.");
+  }
   let findID;
   try {
     findID = new ObjectId(req.params._id);
@@ -1815,7 +1827,10 @@ app.get("/api/Lecture", loginCheck, permissionCheck(Role("manager"),Role("admin"
   }
 });
 
-app.post("/api/Lecture", loginCheck, permissionCheck(Role("manager"),Role("admin")), async (req, res) => {
+app.post("/api/Lecture", loginCheck, async (req, res) => {
+  if (req["user"]["ID"] === "guest") {
+    return res.send("게스트 계정은 저장, 수정, 삭제가 불가능합니다.");
+  }
   //이 코드 부분에서 강의 추가 시 강의에서 사용하는 교재를 TextbookOfLecture에 추가하도록 수정 필요
   //>>transaction사용: (insertone>Lecture, insertMany>TextbookOfLecture)
   const newLecture = req.body;
@@ -1867,7 +1882,10 @@ app.get("/api/Lecture/:lectureid", loginCheck, permissionCheck(Role("manager"),R
   });
 });
 
-app.put("/api/Lecture", loginCheck, permissionCheck(Role("manager"),Role("admin")), (req, res) => {
+app.put("/api/Lecture", loginCheck, (req, res) => {
+  if (req["user"]["ID"] === "guest") {
+    return res.send("게스트 계정은 저장, 수정, 삭제가 불가능합니다.");
+  }
   const newLecture = req.body;
   let findID;
   try {
@@ -1885,7 +1903,10 @@ app.put("/api/Lecture", loginCheck, permissionCheck(Role("manager"),Role("admin"
   });
 });
 
-app.delete("/api/Lecture/:lectureid", loginCheck, permissionCheck(Role("manager"),Role("admin")), (req, res) => {
+app.delete("/api/Lecture/:lectureid", loginCheck, (req, res) => {
+  if (req["user"]["ID"] === "guest") {
+    return res.send("게스트 계정은 저장, 수정, 삭제가 불가능합니다.");
+  }
   const paramID = decodeURIComponent(req.params.lectureid);
   db.collection("Lecture").deleteOne({ lectureID: paramID }, (err, result) => {
     if (err) {
@@ -2546,7 +2567,10 @@ app.get("/api/TRnow", loginCheck, permissionCheck(Role("manager"),Role("admin"))
 });
 
 // Weeklymeeting 관련 코드
-app.post("/api/Weeklymeeting/:date", loginCheck, permissionCheck(Role("manager"),Role("admin")), (req, res) => {
+app.post("/api/Weeklymeeting/:date", loginCheck, (req, res) => {
+  if (req["user"]["ID"] === "guest") {
+    return res.send("게스트 계정은 저장, 수정, 삭제가 불가능합니다.");
+  }
   const newWeeklymeeting = req.body;
   const paramDate = decodeURIComponent(req.params.date);
   db.collection("Weeklymeeting").findOne({ 회의일: paramDate }, (err, result) => {
@@ -2575,7 +2599,10 @@ app.get("/api/Weeklymeeting/:date", loginCheck, permissionCheck(Role("manager"),
   });
 });
 
-app.put("/api/Weeklymeeting/:date", loginCheck, permissionCheck(Role("manager"),Role("admin")), (req, res) => {
+app.put("/api/Weeklymeeting/:date", loginCheck, (req, res) => {
+  if (req["user"]["ID"] === "guest") {
+    return res.send("게스트 계정은 저장, 수정, 삭제가 불가능합니다.");
+  }
   const newWeeklymeeting = req.body;
   const paramDate = decodeURIComponent(req.params.date);
   delete newWeeklymeeting["_id"];
@@ -2587,7 +2614,10 @@ app.put("/api/Weeklymeeting/:date", loginCheck, permissionCheck(Role("manager"),
   });
 });
 
-app.delete("/api/Weeklymeeting/:date", loginCheck, permissionCheck(Role("manager"),Role("admin")), (req, res) => {
+app.delete("/api/Weeklymeeting/:date", loginCheck, (req, res) => {
+  if (req["user"]["ID"] === "guest") {
+    return res.send("게스트 계정은 저장, 수정, 삭제가 불가능합니다.");
+  }
   const paramDate = decodeURIComponent(req.params.date);
   db.collection("Weeklymeeting").deleteOne({ 회의일: paramDate }, (err, result) => {
     if (err) {
@@ -2598,7 +2628,10 @@ app.delete("/api/Weeklymeeting/:date", loginCheck, permissionCheck(Role("manager
 });
 
 // Weeklystudyfeedback 관련 코드
-app.post("/api/Weeklystudyfeedback/:ID/:feedbackDate", loginCheck, permissionCheck(Role("manager"),Role("admin")), (req, res) => {
+app.post("/api/Weeklystudyfeedback/:ID/:feedbackDate", loginCheck, (req, res) => {
+  if (req["user"]["ID"] === "guest") {
+    return res.send("게스트 계정은 저장, 수정, 삭제가 불가능합니다.");
+  }
   const newWeeklystudyfeedback = req.body;
   const paramDate = decodeURIComponent(req.params.feedbackDate);
   const ID = decodeURIComponent(req.params.ID);
@@ -2629,7 +2662,10 @@ app.get("/api/Weeklystudyfeedback/:ID/:feedbackDate", loginCheck, permissionChec
   });
 });
 
-app.put("/api/Weeklystudyfeedback/:ID/:feedbackDate", loginCheck, permissionCheck(Role("manager"),Role("admin")), (req, res) => {
+app.put("/api/Weeklystudyfeedback/:ID/:feedbackDate", loginCheck, (req, res) => {
+  if (req["user"]["ID"] === "guest") {
+    return res.send("게스트 계정은 저장, 수정, 삭제가 불가능합니다.");
+  }
   const newWeeklystudyfeedback = req.body;
   const paramDate = decodeURIComponent(req.params.feedbackDate);
   const ID = decodeURIComponent(req.params.ID);
@@ -2642,7 +2678,10 @@ app.put("/api/Weeklystudyfeedback/:ID/:feedbackDate", loginCheck, permissionChec
   });
 });
 
-app.delete("/api/Weeklystudyfeedback/:ID/:feedbackDate", loginCheck, permissionCheck(Role("manager"),Role("admin")), (req, res) => {
+app.delete("/api/Weeklystudyfeedback/:ID/:feedbackDate", loginCheck, (req, res) => {
+  if (req["user"]["ID"] === "guest") {
+    return res.send("게스트 계정은 저장, 수정, 삭제가 불가능합니다.");
+  }
   const paramDate = decodeURIComponent(req.params.feedbackDate);
   const ID = decodeURIComponent(req.params.ID);
   db.collection("WeeklyStudyfeedback").deleteOne({ 학생ID: ID, 피드백일: paramDate }, (err, result) => {
@@ -3102,7 +3141,10 @@ app.get("/api/stickynote", loginCheck, permissionCheck(Role("manager"),Role("adm
       });
 });
 
-app.post("/api/stickynote", loginCheck, permissionCheck(Role("manager"),Role("admin")), function (req, res) {
+app.post("/api/stickynote", loginCheck, function (req, res) {
+  if (req["user"]["ID"] === "guest") {
+    return res.send("게스트 계정은 저장, 수정, 삭제가 불가능합니다.");
+  }
   const newStickynote = req.body;
   // console.log(req.body);
   db.collection("stickynote").findOne({ note: newStickynote["note"] }, function (err, result) {
@@ -3121,7 +3163,10 @@ app.post("/api/stickynote", loginCheck, permissionCheck(Role("manager"),Role("ad
   });
 });
 
-app.put("/api/stickynote/:id", loginCheck, permissionCheck(Role("manager"),Role("admin")), (req, res) => {
+app.put("/api/stickynote/:id", loginCheck, (req, res) => {
+  if (req["user"]["ID"] === "guest") {
+    return res.send("게스트 계정은 저장, 수정, 삭제가 불가능합니다.");
+  }
   const newstickynote = req.body;
   let findID;
   try {
